pyproject.toml
[tool.poetry]
name        = "open-memory-suite"
version     = "0.1.0"
description = "Benchmark + frugal‑memory dispatcher for LLM long‑term memory"
authors     = ["Open Memory Suite Team"]
readme      = "README.md"
packages    = [{include = "open_memory_suite"}]

[tool.poetry.dependencies]
python               = "^3.10"          # 3.10, 3.11, 3.12 all OK
torch                = "^2.4.0"         # 3.12 wheels supported
transformers         = "^4.54.0"        # ≥ 4.54 dropped the pydantic <2 pin
sentence-transformers = "^2.6.0"
faiss-cpu            = "^1.9.0"         # 3.12 fix merged upstream
numpy                = "^1.26.4"
pandas               = "^2.2.0"
datasets             = "^2.19.0"
streamlit            = "^1.35.0"        # Works on Py 3.12
langchain            = "^0.3.3"         # 0.3+ requires pydantic 2 
pydantic             = "^2.7.1"         # Single source of truth
peft                = "^0.8.2"
accelerate          = "^0.29.0"         # runtime for PEFT
huggingface-hub     = "^0.34.0"         # explicit pin
wandb               = "^0.18.0"
pyyaml              = "^6.0.1"
requests            = "^2.32.3"
aiohttp             = "^3.9.5"
tqdm                = "^4.66.4"
aiofiles            = "^23.2.0"
rich                = "^13.7.0"         # Beautiful CLI output
click               = "^8.1.7"          # CLI framework
tiktoken            = "^0.8.0"          # Tokenizer for cost estimation
scikit-learn        = "^1.5.0"         # ML utilities for training
fastapi             = "^0.115.0"       # REST API server
uvicorn             = "^0.32.0"         # ASGI server for FastAPI
xgboost             = "^2.1.0"          # XGBoost for ML routing classifier

[tool.poetry.group.dev.dependencies]
pytest         = "^8.2.0"
pytest-asyncio = "^0.24.0"
pytest-timeout = "^2.3.1"
black          = "^24.7.0"
ruff           = "^0.4.6"
mypy           = "^1.10.0"
coverage       = "^7.5.3"
pytest-cov     = "^5.0.0"

[tool.poetry.group.optional.dependencies]
zep-python = "^2.0.0"
neo4j      = "^5.20.0"
vllm = {version = "^0.10.0", optional = true, python = "<3.13"}
spacy = {version = "^3.7.0", optional = true, python = ">=3.9,<3.13"}

[build-system]
requires      = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length    = 88
target-version = ["py310"]

[tool.ruff]
line-length    = 88
target-version = "py310"
select         = ["E", "F", "W", "C90", "I"]

[tool.mypy]
python_version     = "3.10"
warn_return_any    = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths        = ["tests"]
addopts          = "--cov=open_memory_suite --cov-report=html --cov-report=term-missing --timeout=1200"
asyncio_mode     = "auto"
timeout          = 1200  # 20 minutes for ML training tests